# api/openapi.yaml
openapi: 3.1.0
info:
  title: Runtrue API
  version: 0.1.0
  description: >-
    Preliminary control-plane contract. Mutation responses may include
    Idempotency-Replayed. Secret values and enrollment bearer tokens are never
    returned by read endpoints. Enrolled-runner secret and OIDC broker RPCs
    use the separately versioned `runtrue.runner.v1.RunnerControl` protobuf
    contract; they require mTLS plus a live certificate-owned Open session and
    are intentionally not modeled as HTTP endpoints here.
servers:
  - url: https://ci.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/capsules/{capsule_id}/runs:
    post:
      operationId: createRun
      parameters:
        - {$ref: '#/components/parameters/CapsuleId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRunRequest'}
      responses:
        '201':
          description: Authorized run and jobs durably created
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/v1/runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CancelRunRequest'}
      responses:
        '202':
          description: Cancellation durably requested
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/runs/{run_id}/replay-bundle:
    post:
      operationId: createReplayBundle
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses:
        '201':
          description: Canonical replay bundle sealed and stored
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReplayBundleMetadata'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      operationId: getReplayBundle
      parameters:
        - {$ref: '#/components/parameters/RunId'}
      responses:
        '200':
          description: Canonical sealed replay envelope
          content:
            application/vnd.runtrue.replay+json:
              schema: {type: object}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    browserAccess:
      type: apiKey
      in: cookie
      name: runtrue_access
    browserRefresh:
      type: apiKey
      in: cookie
      name: runtrue_refresh
    browserCsrf:
      type: apiKey
      in: cookie
      name: runtrue_csrf
...
  headers:
    IdempotencyReplayed:
      description: Whether this response represents an earlier durable mutation
      schema: {type: string, enum: ['true', 'false']}
  parameters:
...
    CapsuleId:
      name: capsule_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
    RunId:
      name: run_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional mutation key scoped to the operation and resource
      schema: {type: string, minLength: 1, maxLength: 200}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: Bearer or webhook authentication failed
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: Idempotency subject, immutable state, approval subject, lease, or fence conflict
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 8192
    Digest:
      type: string
      pattern: '^sha256:[0-9a-f]{64}$'
...
    CreateRunRequest:
      type: object
      additionalProperties: false
      properties:
        priority: {type: integer, format: int32, default: 0}
        source_snapshot_id:
          allOf: [{$ref: '#/components/schemas/Identifier'}]
          description: Required for a remote capsule whose signed context contains a source-tree digest; rejected for an unbound capsule.
...
    CancelRunRequest:
      type: object
      additionalProperties: false
      properties:
        reason: {type: string, maxLength: 8192}
    Run:
      type: object
      additionalProperties: false
      required: [id, capsule_id, status, created_at, started_at, completed_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        capsule_id: {$ref: '#/components/schemas/Identifier'}
        status: {type: string}
        created_at: {type: string, format: date-time}
        started_at: {type: ['string', 'null'], format: date-time}
        completed_at: {type: ['string', 'null'], format: date-time}
...
    ReplayBundleMetadata:
      type: object
      additionalProperties: false
      required: [id, digest, run_id, download_url, expires_at]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        digest: {$ref: '#/components/schemas/Digest'}
        run_id: {$ref: '#/components/schemas/Identifier'}
        download_url: {type: string, format: uri-reference}
        expires_at: {type: string, format: date-time}
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, request_id]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        request_id: {type: string}